Only extract remote address from peer SDP
We need to extract the remote address of the client to pass along to the
bridge. This adds an extra check to see whether the retrieved address is
a remote address (as opposed to local, loopback, or unspecified). It
also expands the SDP regexes to check for addresses in the ice
candidates.

This fixes issue #16