Do not set NAT type to unknown after failed test
If a previous NAT check has succeeded and we fail to contact the probe
server, do not set the NAT type to unknown. Most proxies will probably
not change their NAT type frequently, so this will prevent an outage of
the probe service from wiping out our unrestricted proxy pool.